@@ -25,98 +23,101 @@
   "author": "Sandrino Di Mattia",
   "license": "MIT",
   "dependencies": {
-    "async": "~1.5.2",
+    "async": "^1.5.2",
     "auth0": "^2.1.0",
-    "auth0-oauth2-express": "^1.1.7",
-    "aws-sdk": "^2.2.47",
+    "auth0-oauth2-express": "^1.1.8",
+    "aws-sdk": "^2.3.17",
     "axios": "^0.9.1",
     "babel": "^6.5.2",
-    "babel-core": "^6.7.4",
-    "babel-plugin-transform-runtime": "^6.6.0",
-    "babel-polyfill": "^6.7.4",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-core": "^6.9.1",
+    "babel-plugin-transform-runtime": "^6.9.0",
+    "babel-polyfill": "^6.9.1",
+    "babel-preset-es2015": "^6.9.0",
     "babel-preset-es2015-loose": "^7.0.0",
     "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.7.2",
-    "babel-runtime": "^6.6.1",
-    "bluebird": "^3.3.4",
-    "body-parser": "^1.15.0",
-    "ejs": "^2.4.1",
+    "babel-register": "^6.9.0",
+    "babel-runtime": "^6.9.2",
+    "bluebird": "^3.4.0",
+    "body-parser": "^1.15.1",
+    "ejs": "^2.4.2",
     "express": "^4.13.4",
-    "express-jwt": "^3.3.0",
+    "express-jwt": "^3.4.0",
     "glob": "^7.0.3",
     "json-loader": "^0.5.4",
-    "lodash": "^3.9.3",
+    "jsonwebtoken": "^7.0.0",
+    "lodash": "^3.10.1",
     "lowdb": "^0.12.5",
-    "lru-memoizer": "^1.1.0",
-    "moment": "^2.12.0",
+    "lru-memoizer": "^1.5.0",
+    "moment": "^2.13.0",
     "mongo-getdb": "^2.0.0",
     "mongodb-uri": "^0.9.7",
     "morgan": "^1.7.0",
+    "ms": "^0.7.1",
     "nconf": "^0.8.4",
     "node-uuid": "^1.4.7",
-    "request": "^2.69.0",
+    "request": "^2.72.0",
     "semver": "^5.1.0",
     "validate.js": "^0.9.0",
-    "webtask-tools": "^1.3.0",
+    "webtask-tools": "^1.4.0",
     "winston": "^2.2.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.3.4",
-    "babel-eslint": "^6.0.0",
+    "autoprefixer": "^6.3.6",
+    "babel-eslint": "^6.0.4",
     "babel-loader": "^6.2.4",
     "babel-preset-react": "^6.5.0",
     "babel-preset-react-hmre": "^1.1.1",
     "chai": "^3.5.0",
-    "classnames": "^2.2.3",
+    "classnames": "^2.2.5",
     "css-loader": "^0.23.1",
-    "eslint": "^2.4.0",
+    "eslint": "^2.11.1",
     "eslint-config-airbnb": "^6.2.0",
-    "eslint-plugin-babel": "^3.1.0",
-    "eslint-plugin-react": "^4.2.3",
+    "eslint-plugin-babel": "^3.2.0",
+    "eslint-plugin-react": "^4.3.0",
     "exports-loader": "^0.6.3",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
-    "history": "^2.0.1",
-    "immutable": "^3.7.6",
+    "history": "^2.1.2",
+    "immutable": "^3.8.1",
     "imports-loader": "^0.6.5",
     "json-loader": "^0.5.4",
     "jwt-decode": "^2.0.1",
-    "mocha": "^2.4.5",
-    "nodemon": "^1.9.1",
+    "mocha": "^2.5.3",
+    "nock": "^8.0.0",
+    "nodemon": "^1.9.2",
     "postcss-focus": "^1.0.0",
     "postcss-loader": "^0.8.2",
     "postcss-reporter": "^1.3.3",
     "postcss-simple-vars": "^1.2.0",
     "raw-loader": "^0.5.1",
-    "react": "^0.14.7",
-    "react-bootstrap": "^0.28.4",
-    "react-dom": "^0.14.7",
-    "react-loader-advanced": "^1.1.2",
+    "react": "^0.14.8",
+    "react-bootstrap": "^0.28.5",
+    "react-dom": "^0.14.8",
+    "react-loader-advanced": "^1.4.0",
     "react-pure-render": "^1.0.2",
-    "react-redux": "^4.4.1",
-    "react-router": "^2.0.1",
-    "react-router-redux": "^4.0.0",
+    "react-redux": "^4.4.5",
+    "react-router": "^2.4.1",
+    "react-router-redux": "^4.0.5",
     "react-transform-hmr": "^1.0.4",
-    "redux": "^3.3.1",
-    "redux-devtools": "^3.1.1",
-    "redux-devtools-dock-monitor": "^1.1.0",
-    "redux-devtools-log-monitor": "^1.0.5",
+    "redux": "^3.5.2",
+    "redux-devtools": "^3.3.1",
+    "redux-devtools-dock-monitor": "^1.1.1",
+    "redux-devtools-log-monitor": "^1.0.11",
     "redux-form": "^4.2.2",
     "redux-logger": "^2.6.1",
-    "redux-promise-middleware": "^2.3.3",
+    "redux-promise-middleware": "^2.4.0",
     "redux-simple-router": "^2.0.4",
-    "redux-thunk": "^2.0.1",
+    "redux-thunk": "^2.1.0",
     "request-promise": "^2.0.1",
     "rimraf": "^2.5.2",
     "style-loader": "^0.13.1",
     "url-loader": "^0.5.7",
-    "webpack": "^1.12.14",
-    "webpack-dev-middleware": "^1.6.0",
+    "webpack": "^1.13.1",
+    "webpack-dev-middleware": "^1.6.1",
     "webpack-dev-server": "^1.14.1",
     "webpack-hot-middleware": "^2.10.0",
-    "webpack-sources": "^0.1.1",
+    "webpack-sources": "^0.1.2",
     "webpack-stats-plugin": "^0.1.1",
-    "webtask-bundle": "^1.3.2"
+    "webtask-bundle": "^1.3.3"
   }
 }
