@@ -55,13 +55,7 @@ export function loadCredentials() {
           localStorage.setItem('apiToken', hash.id_token);
           localStorage.setItem('userProfile', JSON.stringify(profile));
 
-          dispatch({
-            type: constants.LOGIN_SUCCESS,
-            payload: {
-              id_token: hash.id_token,
-              user: profile
-            }
-          });
+          setToken(dispatch, hash.id_token, profile);
 
           if (hash.state) {
             dispatch(push(hash.state));
@@ -75,62 +69,52 @@ export function loadCredentials() {
     const id_token = localStorage.getItem('apiToken');
     const profile = localStorage.getItem('userProfile');
     if (id_token && profile) {
-      const decodedToken = jwtDecode(id_token);
-      if (isExpired(decodedToken)) {
-        return;
-      }
-
-      axios.defaults.headers.common['Authorization'] = `Bearer ${id_token}`;
-
-      dispatch({
-        type: constants.LOADED_TOKEN,
-        payload: {
-          id_token
-        }
-      });
-
-      dispatch({
-        type: constants.LOGIN_SUCCESS,
-        payload: {
-          id_token,
-          user: JSON.parse(profile)
-        }
-      });
+      setToken(dispatch, id_token, JSON.parse(profile));
       return;
     }
 
     // Webtask support.
     const apiToken = sessionStorage.getItem('apiToken');
     if (apiToken) {
-      const decodedToken = jwtDecode(apiToken);
-      axios.defaults.headers.common['Authorization'] = `Bearer ${apiToken}`;
-
-      dispatch({
-        type: constants.LOADED_TOKEN,
-        payload: {
-          apiToken
-        }
-      });
-
-      dispatch({
-        type: constants.LOGIN_SUCCESS,
-        payload: {
-          apiToken,
-          user: decodedToken
-        }
-      });
+      setToken(dispatch, apiToken);
       return;
     }
   };
 }
 
 function isExpired(decodedToken) {
-  if(typeof decodedToken.exp === 'undefined') {
+  if (typeof decodedToken.exp === 'undefined') {
     return true;
   }
 
-  let d = new Date(0);
+  const d = new Date(0);
   d.setUTCSeconds(decodedToken.exp);
 
   return !(d.valueOf() > (new Date().valueOf() + (1000)));
+}
+
+
+function setToken(dispatch, token, profile) {
+  const decodedToken = jwtDecode(token);
+  if (isExpired(decodedToken)) {
+    return;
+  }
+
+  axios.defaults.headers.common.Authorization = `Bearer ${token}`;
+
+  dispatch({
+    type: constants.LOADED_TOKEN,
+    payload: {
+      token
+    }
+  });
+
+  dispatch({
+    type: constants.LOGIN_SUCCESS,
+    payload: {
+      token,
+      decodedToken,
+      user: profile || decodedToken
+    }
+  });
 }
