@@ -3,11 +3,13 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
+import * as constants from './constants';
 import { useRouterHistory } from 'react-router'
 import { createHistory } from 'history'
 import { routerMiddleware, syncHistoryWithStore } from 'react-router-redux';
 
 import { loadCredentials } from './actions/auth';
+import { fetchRuleStatus } from './actions/configuration';
 import routes from './routes';
 import configureStore from './store/configureStore';
 
@@ -22,7 +24,18 @@ const history = useRouterHistory(createHistory)({
 const store = configureStore([ routerMiddleware(history) ], { });
 const reduxHistory = syncHistoryWithStore(history, store);
 
-// Fire first events.
+// Check if the rule is enabled.
+store.subscribe(() => {
+  switch (store.getState().lastAction.type) {
+    case constants.LOGIN_SUCCESS:
+    case constants.SAVE_CONFIGURATION_REJECTED:
+    case constants.SAVE_CONFIGURATION_FULFILLED:
+      store.dispatch(fetchRuleStatus());
+      break;
+    default:
+      break;
+  }
+});
 store.dispatch(loadCredentials());
 
 // Render application.
