@@ -20,10 +20,10 @@ class RoleForm extends Component {
       <Modal.Footer>
         <ButtonToolbar>
           <Button bsSize="small" disabled={ loading || submitting } onClick={this.props.onClose}>
-            Cancel
+            <i className="icon icon-budicon-501"></i> Cancel
           </Button>
           <Button bsStyle="primary" bsSize="small" disabled={ loading || submitting } onClick={handleSubmit}>
-            Save
+            <i className="icon icon-budicon-245"></i> Save
           </Button>
         </ButtonToolbar>
       </Modal.Footer>
