@@ -1,25 +1,43 @@
-import { Component } from 'react';
+import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
 import { logout } from '../actions/auth';
+import { configurationActions } from '../actions';
 
-import DevTools from './DevTools';
 import Header from '../components/Header';
+import RuleStatus from '../components/Configuration/RuleStatus';
 import { Sidebar, SidebarItem } from '../components/Dashboard';
 
 class App extends Component {
+  static propTypes = {
+    user: PropTypes.object,
+    issuer: PropTypes.string,
+    ruleStatus: PropTypes.object,
+    logout: PropTypes.func,
+    goToRules: PropTypes.func,
+    goToConfiguration: PropTypes.func
+  };
+
   render() {
     return (
       <div>
-        <Header user={this.props.user} onLogout={this.props.logout} />
+        <Header user={this.props.user} issuer={this.props.issuer} onLogout={this.props.logout} />
         <div className="container">
           <div className="row">
             <Sidebar>
               <SidebarItem title="Applications" route="/applications" icon="icon icon-budicon-375" />
               <SidebarItem title="Groups" route="/groups" icon="icon icon-budicon-322" />
               <SidebarItem title="Users" route="/users" icon="icon icon-budicon-292" />
+              <SidebarItem title="Configuration" route="/configuration/" icon="icon icon-budicon-329">
+                <SidebarItem title="Rule" route="/configuration/rule" icon="icon icon-budicon-519" />
+                <SidebarItem title="Access" route="/configuration/access" icon="icon icon-budicon-519" />
+              </SidebarItem>
             </Sidebar>
             <div id="content" className="col-xs-10">
+              <RuleStatus ruleStatus={this.props.ruleStatus}
+                goToConfiguration={this.props.goToConfiguration}
+                goToRules={this.props.goToRules}
+              />
               { this.props.children }
             </div>
           </div>
@@ -31,8 +49,10 @@ class App extends Component {
 
 function select(state) {
   return {
-    user: state.auth.get('user')
+    issuer: state.auth.get('issuer'),
+    user: state.auth.get('user'),
+    ruleStatus: state.ruleStatus
   };
 }
 
-export default connect(select, { logout })(App);
+export default connect(select, { logout, ...configurationActions })(App);
