@@ -1,3 +1,4 @@
+import url from 'url';
 import { fromJS } from 'immutable';
 
 import * as constants from '../constants';
@@ -7,7 +8,9 @@ const initialState = {
   error: null,
   isAuthenticated: false,
   isAuthenticating: false,
+  issuer: null,
   token: null,
+  decodedToken: null,
   user: null
 };
 
@@ -27,12 +30,15 @@ export const auth = createReducer(fromJS(initialState), {
       isAuthenticated: true,
       isAuthenticating: false,
       user: action.payload.user,
-      token: action.payload.id_token
+      token: action.payload.token,
+      decodedToken: action.payload.decodedToken,
+      issuer: url.parse(action.payload.decodedToken.iss).hostname
     }),
   [constants.LOGOUT_SUCCESS]: (state) =>
     state.merge({
       user: null,
       token: null,
+      decodedToken: null,
       isAuthenticated: false
     })
 });
