@@ -1,5 +1,6 @@
 export authenticate from './authenticate';
 export authenticateOrApiKey from './authenticateOrApiKey';
 export errorHandler from './errorHandler';
-export readSubject from './readSubject';
 export managementClient from './managementClient';
+export readSubject from './readSubject';
+export validateHookToken from './validateHookToken';
