@@ -1,7 +1,4 @@
-import _ from 'lodash';
-import Promise from 'bluebird';
 import { Router } from 'express';
-
 import { isApplicationAccessAllowed, getUserGroups } from '../lib/queries';
 
 export default (db) => {
@@ -16,7 +13,10 @@ export default (db) => {
 
     getUserGroups(db, userId, connectionName, groups)
       .then((userGroups) => isApplicationAccessAllowed(db, clientId, userGroups)
-        .then(isAccessAllowed => res.json({ groups: userGroups.map((group) => group.name), accessGranted: isAccessAllowed }))
+        .then(isAccessAllowed => res.json({
+          groups: userGroups.map((group) => group.name),
+          accessGranted: isAccessAllowed
+        }))
       )
       .catch(next);
   });
