@@ -7,5 +7,6 @@ nconf
   .file(path.join(__dirname, '../server/config.json'))
   .defaults({
     NODE_ENV: 'test',
-    DATA_CACHE_MAX_AGE: 0
+    DATA_CACHE_MAX_AGE: 0,
+    DUMMY_KEY: 'DUMMY_VALUE'
   });
